requirements.txt
tensorflow-cpu
keras
opencv-python-headless
numpy